@@ -14,11 +14,12 @@ class FirestoreService extends DatabaseService {
 
   Stream<List<T>> collectionStream<T>({
     @required String path,
-    @required T builder(Map<String, dynamic> data),
+    @required T builder(Map<String, dynamic> data, String documentId),
   }) {
     final reference = Firestore.instance.collection(path);
     final snapshots = reference.snapshots();
-    return snapshots.map((snapshot) =>
-        snapshot.documents.map((snapshot) => builder(snapshot.data)).toList());
+    return snapshots.map((snapshot) => snapshot.documents
+        .map((snapshot) => builder(snapshot.data, snapshot.documentID))
+        .toList());
   }
 }
