@@ -1,18 +1,20 @@
 import 'package:meta/meta.dart';
 
 class Job {
-  Job({@required this.name, @required this.ratePerHour});
+  Job({@required this.id, @required this.name, @required this.ratePerHour});
 
+  final String id;
   final String name;
   final int ratePerHour;
 
-  factory Job.fromMap(Map<String, dynamic> data) {
+  factory Job.fromMap(Map<String, dynamic> data, String documentId) {
     if (data == null) {
       return null;
     }
     final String name = data['name'];
     final int ratePerHour = data['ratePerHour'];
     return Job(
+      id: documentId,
       name: name,
       ratePerHour: ratePerHour,
     );
@@ -23,5 +25,15 @@ class Job {
       'name': name,
       'ratePerHour': ratePerHour,
     };
+  }
+
+  @override
+  String toString() {
+    return """
+    Job(
+      name: $name,
+      int: $ratePerHour,
+    );
+    """;
   }
 }
